pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "querynest"
version = "1.0.0"
description = "QueryNest MCP MongoDB查询服务"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "QueryNest Team", email = "team@querynest.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "pymongo>=4.0.0",
    "motor>=3.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "structlog>=23.0.0",
    "nltk>=3.8",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[project.scripts]
querynest-mcp = "mcp_server:cli_main"



[project.urls]
Homepage = "https://github.com/querynest/querynest"
Repository = "https://github.com/querynest/querynest.git"
Issues = "https://github.com/querynest/querynest/issues"

[tool.setuptools]
py-modules = ["mcp_server", "config"]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"